--- app/src/main/java/com/nemodream/bangkkujaengi/ParentFragment.kt
@@ -1,31 +1,0 @@
-package com.nemodream.bangkkujaengi
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.nemodream.bangkkujaengi.databinding.FragmentParentBinding
-
-class ParentFragment: Fragment() {
-    private var _binding: FragmentParentBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentParentBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
