+++ app/src/main/res/layout/fragment_customer_container.xml
@@ -9,14 +9,15 @@
         android:id="@+id/customer_container"
         android:layout_width="0dp"
         android:layout_height="0dp"
+        app:layout_constraintBottom_toTopOf="@id/customer_bottom_navigation"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <com.google.android.material.divider.MaterialDivider
+        style="@style/AppDivider"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        style="@style/AppDivider"
         app:layout_constraintBottom_toTopOf="@id/customer_bottom_navigation"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
